Add -known_import flag
Private repositories can take a long time to resolve, and if the repo
host does not properly support go-import meta tags then gazelle may not
be able to resolve them at all.

This change adds a flag that can be specified multiple times to append
arbitrary import paths to the cache of known imports when using the
external import resolver.